\name{Model_evaluation}
\alias{Model_evaluation}
\alias{Model_evaluation}

\title{Evaluate SCE Model Performances}

\description{
  This function facilitates the evaluation of SCE model performances over training,
  out-of-bag validation, and testing datasets, providing a comprehensive summary
  through 22 distinct evaluation metrics.
}

\usage{
Model_evaluation(
  Predictant,
  obs_training,
  obs_testing,
  sim,
  Num_predictor,
  digits = 2
)
}

\arguments{
  \item{Predictant}{
    A string specifying the name of one of the dependent (y) variables to be evaluated
    (e.g., "streamflow"). It is important that the specified name exactly matches
    one of the dependent variable names used in model training.
  }
  \item{obs_training}{
    A data.frame comprising the observations used during the model training phase.
  }
  \item{obs_testing}{
    A data.frame comprising the observations used during the model testing phase.
  }
  \item{sim}{
    A list containing model simulations for training, out-of-bag validation,
    and testing phases. The structure of this list should align with the output
    generated by the model training function.
  }
  \item{Num_predictor}{
    The number of independent (x) variables utilized in the model. This value
    is used to calculate adjusted R-squared metrics and should match the number
    of independent variables specified during model training.
  }
  \item{digits}{
    An integer specifying the number of decimal places to retain when reporting
    evaluation metrics. The default value is 2, which means metrics will be rounded
    to two decimal places.
  }
}

\author{
  Kailong Li <lkl98509509@gmail.com>
}

\examples{
  ## Load SCE package
  library(SCE)

  ## Training data file
  data("Training_input")
  ## Testing data file
  data("Testing_input")

  ## Define independent (x) and dependent (y) variables
  Predictors <- c("Prcp","SRad","Tmax","Tmin","VP","smlt","swvl1","swvl2")
  Predictants <- c("swvl3","swvl4")

  ## Build the SCE model
  Model <- SCE(Training_data = Training_input, X = Predictors, Y = Predictants, mfeature = round(0.5 * length(Predictors)), Nmin = 5, Ntree = 48, alpha = 0.05)

  ## Make predictions
  Results <- Model_simulation(Testing_data = Testing_input, X = Predictors, model = Model)

  ## Evaluate the model over specified predictants
  Evaluation_swvl3 <- Model_evaluation(Predictant = "swvl3", obs_training = Training_input, obs_testing = Testing_input, sim = Results, Num_predictor = length(Predictors), digits = 2)

  Evaluation_swvl4 <- Model_evaluation(Predictant = "swvl4", obs_training = Training_input, obs_testing = Testing_input, sim = Results, Num_predictor = length(Predictors), digits = 2)
}
